refactoring code => adding main to each submodule and moving some code pieces to utils.py
