+++ src/music_sync/apple_music/main.py
@@ -1,0 +1,10 @@
+from music_sync.apple_music.library import write_apple_music_library, prepare_playlists
+
+
+def main():
+    write_apple_music_library()
+    prepare_playlists()
+
+
+if __name__ == "__main__":
+    main()
